requirements.txt
# Core dependencies
requests>=2.28.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Visualization
streamlit>=1.28.0
plotly>=5.18.0
altair>=5.0.0

# Environment management
python-dotenv>=1.0.0
